refactor(add-location): extract the form into a runes component #1134
First step of the in-map move: the details form (captcha, address
prefill, category picker, payments, optional expander, hours accordion,
submit) leaves the page for $components/add-location/AddLocationForm —
born in runes mode per the deferred-conversion decision, since the
map's placement side-panel will host this same component next. The page
becomes a thin legacy host owning chrome, the supertagger column and
the success state (the form reports completion via `onsuccess`). The
shared `methods` array becomes a per-submit local — same payload,
no cross-retry accumulation. No visual or behavioral change; the six
existing add-location e2e specs pass unmodified.
